Refactor scripts

+++ openvpn/openvpn.sh
@@ -1,11 +1,18 @@
-#!/bin/sh
-set -ex
+#!/usr/bin/env bash
+set -euxo pipefail
 
-sudo apt update
-sudo apt install openvpn
-if [ -d ~/openvpn ]; then
-  sudo mv -f /etc/openvpn{,.orig}
-  sudo ln -s ~/openvpn /etc/openvpn
-  sudo systemctl disable systemd-ask-password-wall.path systemd-ask-password-wall.service
-  sudo systemctl stop systemd-ask-password-wall.path systemd-ask-password-wall.service
+export OS=${OSTYPE:-'linux-gnu'}
+export OS_TYPE=`echo ${OS} | tr -d "[:digit:]"`
+
+if [ "${OS_TYPE}" == "linux-gnu" ]; then
+    if [ -f /etc/debian_version ]; then
+        sudo apt update
+        sudo apt install -y openvpn
+        if [ -d ~/openvpn ]; then
+            sudo mv -f /etc/openvpn{,.orig}
+            sudo ln -s ~/openvpn /etc/openvpn
+            sudo systemctl disable systemd-ask-password-wall.path systemd-ask-password-wall.service
+            sudo systemctl stop systemd-ask-password-wall.path systemd-ask-password-wall.service
+        fi
+    fi
 fi
